introduce portfolio calculator factory that will allow different types of performace metrics, for now just TWR is working as before
