+++ apps/api/src/app/portfolio/calculator/portfolio-calculator-test-utils.ts
@@ -1,0 +1,25 @@
+export const symbolProfileDummyData = {
+  activitiesCount: undefined,
+  assetClass: undefined,
+  assetSubClass: undefined,
+  countries: [],
+  createdAt: undefined,
+  id: undefined,
+  sectors: [],
+  updatedAt: undefined
+};
+
+export const activityDummyData = {
+  value: undefined,
+  valueInBaseCurrency: undefined,
+  accountId: undefined,
+  id: undefined,
+  userId: undefined,
+  createdAt: new Date(),
+  updatedAt: new Date(),
+  accountUserId: undefined,
+  comment: undefined,
+  feeInBaseCurrency: undefined,
+  isDraft: false,
+  symbolProfileId: undefined
+};
