+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator-baln-buy-and-sell-in-two-activities.spec.ts
@@ -6,7 +6,14 @@ import { parseDate } from '@ghostfolio/common/helper';
 
 import { Big } from 'big.js';
 
-import { PortfolioCalculator } from './portfolio-calculator';
+import {
+  activityDummyData,
+  symbolProfileDummyData
+} from '../portfolio-calculator-test-utils';
+import {
+  PortfolioCalculatorFactory,
+  PerformanceCalculationType
+} from '../portfolio-calculator.factory';
 
 jest.mock('@ghostfolio/api/app/portfolio/current-rate.service', () => {
   return {
@@ -20,6 +27,7 @@ jest.mock('@ghostfolio/api/app/portfolio/current-rate.service', () => {
 describe('PortfolioCalculator', () => {
   let currentRateService: CurrentRateService;
   let exchangeRateDataService: ExchangeRateDataService;
+  let factory: PortfolioCalculatorFactory;
 
   beforeEach(() => {
     currentRateService = new CurrentRateService(null, null, null, null);
@@ -30,54 +38,66 @@ describe('PortfolioCalculator', () => {
       null,
       null
     );
+
+    factory = new PortfolioCalculatorFactory(
+      currentRateService,
+      exchangeRateDataService
+    );
   });
 
   describe('get current positions', () => {
     it.only('with BALN.SW buy and sell in two activities', async () => {
-      const portfolioCalculator = new PortfolioCalculator({
-        currentRateService,
-        exchangeRateDataService,
-        activities: <Activity[]>[
-          {
-            date: new Date('2021-11-22'),
-            fee: 1.55,
-            quantity: 2,
-            SymbolProfile: {
-              currency: 'CHF',
-              dataSource: 'YAHOO',
-              name: 'Bâloise Holding AG',
-              symbol: 'BALN.SW'
-            },
-            type: 'BUY',
-            unitPrice: 142.9
+      const activities: Activity[] = [
+        {
+          ...activityDummyData,
+          date: new Date('2021-11-22'),
+          fee: 1.55,
+          quantity: 2,
+          SymbolProfile: {
+            ...symbolProfileDummyData,
+            currency: 'CHF',
+            dataSource: 'YAHOO',
+            name: 'Bâloise Holding AG',
+            symbol: 'BALN.SW'
           },
-          {
-            date: new Date('2021-11-30'),
-            fee: 1.65,
-            quantity: 1,
-            SymbolProfile: {
-              currency: 'CHF',
-              dataSource: 'YAHOO',
-              name: 'Bâloise Holding AG',
-              symbol: 'BALN.SW'
-            },
-            type: 'SELL',
-            unitPrice: 136.6
+          type: 'BUY',
+          unitPrice: 142.9
+        },
+        {
+          ...activityDummyData,
+          date: new Date('2021-11-30'),
+          fee: 1.65,
+          quantity: 1,
+          SymbolProfile: {
+            ...symbolProfileDummyData,
+            currency: 'CHF',
+            dataSource: 'YAHOO',
+            name: 'Bâloise Holding AG',
+            symbol: 'BALN.SW'
           },
-          {
-            date: new Date('2021-11-30'),
-            fee: 0,
-            quantity: 1,
-            SymbolProfile: {
-              currency: 'CHF',
-              dataSource: 'YAHOO',
-              name: 'Bâloise Holding AG',
-              symbol: 'BALN.SW'
-            },
-            type: 'SELL',
-            unitPrice: 136.6
-          }
-        ],
+          type: 'SELL',
+          unitPrice: 136.6
+        },
+        {
+          ...activityDummyData,
+          date: new Date('2021-11-30'),
+          fee: 0,
+          quantity: 1,
+          SymbolProfile: {
+            ...symbolProfileDummyData,
+            currency: 'CHF',
+            dataSource: 'YAHOO',
+            name: 'Bâloise Holding AG',
+            symbol: 'BALN.SW'
+          },
+          type: 'SELL',
+          unitPrice: 136.6
+        }
+      ];
+
+      const portfolioCalculator = factory.createCalculator({
+        calculationType: PerformanceCalculationType.TWR,
+        activities,
         currency: 'CHF'
       });
