+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator-novn-buy-and-sell.spec.ts
@@ -6,7 +6,14 @@ import { parseDate } from '@ghostfolio/common/helper';
 
 import { Big } from 'big.js';
 
-import { PortfolioCalculator } from './portfolio-calculator';
+import {
+  activityDummyData,
+  symbolProfileDummyData
+} from '../portfolio-calculator-test-utils';
+import {
+  PerformanceCalculationType,
+  PortfolioCalculatorFactory
+} from '../portfolio-calculator.factory';
 
 jest.mock('@ghostfolio/api/app/portfolio/current-rate.service', () => {
   return {
@@ -20,6 +27,7 @@ jest.mock('@ghostfolio/api/app/portfolio/current-rate.service', () => {
 describe('PortfolioCalculator', () => {
   let currentRateService: CurrentRateService;
   let exchangeRateDataService: ExchangeRateDataService;
+  let factory: PortfolioCalculatorFactory;
 
   beforeEach(() => {
     currentRateService = new CurrentRateService(null, null, null, null);
@@ -30,41 +38,51 @@ describe('PortfolioCalculator', () => {
       null,
       null
     );
+
+    factory = new PortfolioCalculatorFactory(
+      currentRateService,
+      exchangeRateDataService
+    );
   });
 
   describe('get current positions', () => {
     it.only('with NOVN.SW buy and sell', async () => {
-      const portfolioCalculator = new PortfolioCalculator({
-        currentRateService,
-        exchangeRateDataService,
-        activities: <Activity[]>[
-          {
-            date: new Date('2022-03-07'),
-            fee: 0,
-            quantity: 2,
-            SymbolProfile: {
-              currency: 'CHF',
-              dataSource: 'YAHOO',
-              name: 'Novartis AG',
-              symbol: 'NOVN.SW'
-            },
-            type: 'BUY',
-            unitPrice: 75.8
+      const activities: Activity[] = [
+        {
+          ...activityDummyData,
+          date: new Date('2022-03-07'),
+          fee: 0,
+          quantity: 2,
+          SymbolProfile: {
+            ...symbolProfileDummyData,
+            currency: 'CHF',
+            dataSource: 'YAHOO',
+            name: 'Novartis AG',
+            symbol: 'NOVN.SW'
           },
-          {
-            date: new Date('2022-04-08'),
-            fee: 0,
-            quantity: 2,
-            SymbolProfile: {
-              currency: 'CHF',
-              dataSource: 'YAHOO',
-              name: 'Novartis AG',
-              symbol: 'NOVN.SW'
-            },
-            type: 'SELL',
-            unitPrice: 85.73
-          }
-        ],
+          type: 'BUY',
+          unitPrice: 75.8
+        },
+        {
+          ...activityDummyData,
+          date: new Date('2022-04-08'),
+          fee: 0,
+          quantity: 2,
+          SymbolProfile: {
+            ...symbolProfileDummyData,
+            currency: 'CHF',
+            dataSource: 'YAHOO',
+            name: 'Novartis AG',
+            symbol: 'NOVN.SW'
+          },
+          type: 'SELL',
+          unitPrice: 85.73
+        }
+      ];
+
+      const portfolioCalculator = factory.createCalculator({
+        calculationType: PerformanceCalculationType.TWR,
+        activities,
         currency: 'CHF'
       });
